modification for passing the test case

+++ Leetcode/Foursum.java
@@ -21,7 +21,7 @@ public class Foursum{
             int right = nums.length-1;
 
             while(left<right){
-                int sum = nums[i]+nums[mid]+nums[left]+nums[right];
+                long sum = (long)nums[i]+nums[mid]+nums[left]+nums[right];
 
                 if(sum==target){
                     result.add(Arrays.asList(nums[i], nums[mid],nums[left], nums[right]));
